Make metrics a singleton.
