+++ build.gradle
@@ -20,7 +20,7 @@ repositories {
 }
 
 ext {
-    usiVersion= '0.1.17-66eae22-SNAPSHOT'
+    usiVersion= '0.1.17-3b0db2d-SNAPSHOT'
     scalaVersion = '2.12'
     akkaVersion = '2.5.19'
 }
