+++ src/main/java/org/jenkinsci/plugins/mesos/MesosApi.java
@@ -18,8 +18,6 @@ import com.mesosphere.usi.core.models.*;
 import com.mesosphere.usi.core.models.commands.KillPod;
 import com.mesosphere.usi.core.models.commands.LaunchPod;
 import com.mesosphere.usi.core.models.commands.SchedulerCommand;
-import com.mesosphere.usi.metrics.dropwizard.conf.HistorgramSettings;
-import com.mesosphere.usi.metrics.dropwizard.conf.MetricsSettings;
 import com.mesosphere.usi.repository.PodRecordRepository;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
@@ -36,14 +34,12 @@ import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
 import javax.annotation.Nonnull;
-import jenkins.metrics.api.Metrics;
 import jenkins.model.Jenkins;
 import org.apache.mesos.v1.Protos;
 import org.jenkinsci.plugins.mesos.MesosCloud.DcosAuthorization;
 import org.jenkinsci.plugins.mesos.api.Settings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import scala.Option;
 import scala.compat.java8.OptionConverters;
 import scala.concurrent.ExecutionContext;
 
@@ -125,25 +121,9 @@ public class MesosApi {
     this.materializer = ActorMaterializer.create(system);
     this.context = system.dispatcher();
 
-    // Construct metrics
-    MetricsSettings metricsSettings =
-        new MetricsSettings(
-            sanitize(frameworkName),
-            HistorgramSettings.apply(
-                HistorgramSettings.apply$default$1(),
-                HistorgramSettings.apply$default$2(),
-                HistorgramSettings.apply$default$3(),
-                HistorgramSettings.apply$default$4(),
-                HistorgramSettings.apply$default$5()),
-            Option.empty(),
-            Option.empty());
-    final com.mesosphere.usi.metrics.Metrics metrics =
-        new com.mesosphere.usi.metrics.dropwizard.DropwizardMetrics(
-            metricsSettings, Metrics.metricRegistry());
-
     URL masterUrl =
         MasterDetector$.MODULE$
-            .apply(master, metrics)
+            .apply(master, Metrics.getInstance(frameworkName))
             .getMaster(context)
             .toCompletableFuture()
             .get();
@@ -181,7 +161,9 @@ public class MesosApi {
     commands =
         connectClient(clientSettings, provider)
             .thenCompose(
-                client -> Scheduler.fromClient(client, repository, metrics, schedulerSettings))
+                client ->
+                    Scheduler.fromClient(
+                        client, repository, Metrics.getInstance(frameworkName), schedulerSettings))
             .thenApply(builder -> runScheduler(builder.getFlow(), materializer))
             .get();
 
@@ -383,11 +365,6 @@ public class MesosApi {
         stateMap.remove(podStateEvent.id());
       }
     }
-  }
-
-  /** @return a santized prefix for Dropwizard metrics. */
-  public static String sanitize(String prefix) {
-    return prefix.replaceAll("[^a-zA-Z0-9\\-\\.]", "-");
   }
 
   /** test method to set the agent timeout duration */
